pyproject.toml
[tool.setuptools.packages.find]
where = ["genai_workshop"]

[project]
name = "genaiworkshop"
version = "0.1.0"
description = "GenAI workshop to explore different implementations potentially used for StackAdapt"
readme = "README.md"
requires-python = "==3.10.9"
dependencies = [
    'rich',
    'python-dotenv',
    'openai',
    'requests',
    'pillow',
    'google-genai',
    'torch',
    'torchvision',
    'transformers',
    'diffusers',
    'numpy',
    'kornia',
    'timm',
    "stackadapt.py-gpt-service @ file:///Users/ivan.company/dev/stackadapt/py-gpt-service"
]
